requirements-backend.txt
fastapi==0.135.3
uvicorn==0.44.0
psycopg2-binary==2.9.11
pydantic==2.13.0
python-dotenv==1.2.2
requests==2.33.1
pytest==9.0.3